feat: add compact and expanded list views for books
This commit introduces two new components, CompactListContent and ExpandedListContent, to provide users with alternative views of their book lists. The CompactListContent component offers a summarized view, ideal for quick glances, while the ExpandedListContent component provides a detailed view, including options to edit, delete, or export book entries. Additionally, modifications have been made to the BookDetailsForm, ListOfBooks, and SearchBooks components to integrate these new views seamlessly and enhance the overall user experience.
