+++ src/components/ListOfBooks.js
@@ -1,28 +1,14 @@
-import React, { useRef } from "react";
-import {
-  Button,
-  Heading,
-  VStack,
-  StackDivider,
-  Text,
-  IconButton,
-  Tooltip,
-  HStack,
-  Box,
-  AlertDialog,
-  AlertDialogBody,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogContent,
-  AlertDialogOverlay,
-  useDisclosure
-} from "@chakra-ui/react";
+import React, { useState } from "react";
+import { VStack } from "@chakra-ui/react";
+import CompactListContent from './CompactListContent';
+import ExpandedListContent from './ExpandedListContent';
 
-import { DeleteIcon, EditIcon } from "@chakra-ui/icons";
+
 
 const ListOfBooks = ({ booksList, onDeleteBook, onEditBook, onClearBooksList }) => {
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const cancelRef = useRef();
+  const [isExpanded, setIsExpanded] = useState(false);
+
+  const handleToggleExpand = () => setIsExpanded(!isExpanded)
   
   const handleExport = () => {
     const headers =
@@ -58,122 +44,38 @@ const ListOfBooks = ({ booksList, onDeleteBook, onEditBook, onClearBooksList }) 
     URL.revokeObjectURL(url);
   };
 
-  const reversedBooksList = [...booksList].reverse();
+  
 
   return (
-    <VStack
-      spacing={4}
-      align="stretch"
-      borderWidth="1px"
-      borderRadius="lg"
-      borderColor="gray.400"
-      flex="1"
-      p={4}
-      w="20vw"
-      backgroundColor="white"
-    >
-      <Heading as="h3" size="md">
-        Lista de libros {booksList.length > 0 && `(${booksList.length})`}
-      </Heading>
-      <HStack>
-        <Button
-          flex="3"
-          onClick={handleExport}
-          colorScheme="green"
-          isDisabled={booksList.length === 0}
-        >
-          Exportar como .TAB
-        </Button>
-        <Button flex="1" colorScheme='red' variant='outline' onClick={onOpen} isDisabled={booksList.length === 0}>
-        Vaciar lista
-      </Button>
-
-        <AlertDialog
-        isOpen={isOpen}
-        leastDestructiveRef={cancelRef}
-        onClose={onClose}
-      >
-        <AlertDialogOverlay>
-          <AlertDialogContent>
-            <AlertDialogHeader fontSize='lg' fontWeight='bold'>
-              Vaciar lista de libros
-            </AlertDialogHeader>
-
-            <AlertDialogBody>
-              ¿Estás seguro? Esta acción no se puede deshacer.
-            </AlertDialogBody>
-
-            <AlertDialogFooter>
-              <Button ref={cancelRef} onClick={onClose}>
-                No, cancelar
-              </Button>
-              <Button colorScheme='red' onClick={() => { onClearBooksList(); onClose(); }} ml={3}>
-                Vaciar
-              </Button>
-            </AlertDialogFooter>
-          </AlertDialogContent>
-        </AlertDialogOverlay>
-      </AlertDialog>
-
-      </HStack>
-      <VStack
-        divider={<StackDivider borderColor="gray.400" />}
-        spacing={4}
-        align="stretch"
-        overflowY="auto"
-        flex="1"
-      >
-        {booksList.length === 0 && (
-          <Text color="gray.500">
-            Agrega libros a la lista para poder exportar
-          </Text>
-        )}
-        {reversedBooksList.map((book, index) => (
-          <Box
-            direction={{ base: "column", sm: "row" }}
-            p={0}
-            variant="alpha"
-            key={index}
-            width="100%"
-          >
-            <HStack>
-              <VStack p={0} flex="1" alignItems="flex-start">
-                <Heading as="h4" size="sm">
-                  {book.title} - {book.author}
-                </Heading>
-                <Text fontSize="sm">
-                  {book.book_id || "Sin Nº de articulo"}
-                </Text>
-              </VStack>
-
-              <HStack gap={2}>
-                <Tooltip hasArrow label='Editar' placement='top'>
-                  <IconButton
-                    size="sm"
-                    variant='outline'
-                    colorScheme='blue'
-                    aria-label='Editar'
-                    icon={<EditIcon />}
-                    onClick={() => onEditBook(book)}
-                  />
-                </Tooltip>
-                <Tooltip hasArrow label='Borrar' placement='top'>
-                  <IconButton
-                    size="sm"
-                    variant='outline'
-                    colorScheme="red"
-                    aria-label='Borrar'
-                    icon={<DeleteIcon />}
-                    onClick={() => onDeleteBook(book.book_id)}
-                  />
-                </Tooltip>
-              </HStack>
-
-            </HStack>
-          </Box>
-        ))}
-      </VStack>
-    </VStack>
+    <VStack 
+            
+            borderWidth="1px"
+            borderRadius="lg"
+            borderColor="gray.400"
+            p={4}
+            backgroundColor="white"
+            maxW={"340px"}
+            >
+              
+      {isExpanded ? (
+        <ExpandedListContent
+          onToggleExpand={handleToggleExpand}
+          booksList={booksList}
+          onDeleteBook={onDeleteBook}
+          onEditBook={onEditBook}
+          onClearBooksList={onClearBooksList}
+          handleExport={handleExport}
+          booksCount={booksList.length}
+        />
+      ) : (
+        <CompactListContent
+          booksList={booksList}
+          onToggleExpand={handleToggleExpand}
+          booksCount={booksList.length}
+          handleExport={handleExport}
+        />
+      )}
+    </VStack >
   );
 };
